work on message parse and HostRequest building

+++ src/protocol.rs
@@ -73,7 +73,7 @@ pub mod Host {
         operation: ValidOps,
         checksum: u8,         // Wrapping checksum
         size: u8,             // A value between 0 and 4
-        payload: [u8; 4],     // Max payload size over SPI is 4 bytes 
+        payload: [u32; 4],     // Max payload size over SPI is 4 bytes 
 
     }
 
@@ -186,7 +186,7 @@ pub mod Slave {
         proc_id: u8,
         host_config: ValidHostInterfaces,
         size: u8,             // A value between 0 and 4
-        payload: [u8; 4],     // Max payload size over SPI is 4 bytes 
+        payload: [u32; 4],     // Max payload size over SPI is 4 bytes 
 
     }
